#9874: Overload binary add in unary add
