--- docs/source/ttnn/ttnn/ttnn/unary_add_bw.rst
@@ -1,6 +1,0 @@
-.. _ttnn.unary_add_bw:
-
-ttnn.unary_add_bw
-#################
-
-.. autofunction:: ttnn.unary_add_bw
