+++ ttnn/cpp/ttnn/operations/eltwise/unary_backward/device/unary_backward_op.hpp
@@ -18,7 +18,7 @@ enum class UnaryBackwardOpType {
     CLAMP_BW,
     ASSIGN_BW,
     MULTIGAMMALN_BW,
-    UNARY_ADD_BW,
+    ADD_BW,
 };
